Move data layer to SQLite/libSQL for Vercel + Turso prototype hosting
- Drizzle schema and migrations for SQLite; local dev uses .data/local.db, Turso when TURSO_DATABASE_URL is set
- Uploaded files stored in the database (no server disk on Vercel), 4 MB cap
- Resumable demo seeding with a completion marker; db:setup scripts for local and remote
- Docs: Vercel + Turso is prototype-only hosting; production will be scaled up

+++ src/app/(app)/clients/[id]/page.tsx
@@ -83,7 +83,7 @@ export default async function ClientPage({
     outstandingSummary(db, id),
     db.select({ id: users.id, name: users.name }).from(users).where(inArray(users.role, ["sales_exec", "sales_manager", "owner"])),
     db
-      .select({ n: sql<number>`count(*)::int` })
+      .select({ n: sql<number>`count(*)` })
       .from(activities)
       .where(and(eq(activities.clientId, id), inArray(activities.type, ["call", "whatsapp", "email", "meeting"]))),
   ]);
